remove: user_id from job items

+++ app/Http/Controllers/Api/V1/CompanyImagesController.php
@@ -3,22 +3,25 @@
 namespace App\Http\Controllers\Api\V1;
 
 use App\Http\Controllers\Controller;
+use App\Http\Requests\Company\StoreImageRequest;
 use App\Models\Company;
 use App\Models\Image;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CompanyImagesController extends Controller
 {
-    public function store(Request $request, $id)
+    public function store(StoreImageRequest $request, $id)
     {
         if ($request->hasFile('images')) {
-            $request->validate([
-                'images' => 'required',
-                'images.*' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048'
-            ]);
-            $images = $request->file('images');
 
             $company = Company::find($id);
+
+            Gate::authorize('update', $company);
+
+            $validated = $request->validated();
+            $images = $validated['images'];
+
 
             foreach ($images as $key=>$image ) {
                 $imageName = time() . '_' . $key . '.' . $image->extension();
@@ -40,6 +43,10 @@ class CompanyImagesController extends Controller
     public function destroy($id)
     {
         $image = Image::find($id);
+        $company = Company::find($image->company_id);
+
+        Gate::authorize('delete', $company);
+
         $image->delete();
         unlink(public_path('images/companies/'. $image->company_id . '/' . $image->path));
         return response()->json(['message' => 'Image deleted']);
